+++ packages/expo-barcode-scanner/plugin/src/withBarcodeScanner.ts
@@ -1,4 +1,9 @@
-import { AndroidConfig, ConfigPlugin, createRunOncePlugin } from 'expo/config-plugins';
+import {
+  AndroidConfig,
+  ConfigPlugin,
+  createRunOncePlugin,
+  withInfoPlist,
+} from 'expo/config-plugins';
 
 const pkg = require('expo-barcode-scanner/package.json');
 
@@ -11,12 +16,14 @@ const withBarcodeScanner: ConfigPlugin<
     cameraPermission?: string;
   } | void
 > = (config, { microphonePermission, cameraPermission } = {}) => {
-  if (!config.ios) config.ios = {};
-  if (!config.ios.infoPlist) config.ios.infoPlist = {};
-  config.ios.infoPlist.NSCameraUsageDescription =
-    cameraPermission || config.ios.infoPlist.NSCameraUsageDescription || CAMERA_USAGE;
-  config.ios.infoPlist.NSMicrophoneUsageDescription =
-    microphonePermission || config.ios.infoPlist.NSMicrophoneUsageDescription || MICROPHONE_USAGE;
+  withInfoPlist(config, (config) => {
+    config.modResults.NSCameraUsageDescription =
+      cameraPermission || config.modResults.NSCameraUsageDescription || CAMERA_USAGE;
+    config.modResults.NSMicrophoneUsageDescription =
+      microphonePermission || config.modResults.NSMicrophoneUsageDescription || MICROPHONE_USAGE;
+
+    return config;
+  });
 
   return AndroidConfig.Permissions.withPermissions(config, ['android.permission.CAMERA']);
 };