+++ packages/expo-local-authentication/plugin/src/withLocalAuthentication.ts
@@ -1,4 +1,9 @@
-import { AndroidConfig, ConfigPlugin, createRunOncePlugin } from 'expo/config-plugins';
+import {
+  AndroidConfig,
+  ConfigPlugin,
+  createRunOncePlugin,
+  withInfoPlist,
+} from 'expo/config-plugins';
 
 const pkg = require('expo-local-authentication/package.json');
 const FACE_ID_USAGE = 'Allow $(PRODUCT_NAME) to use Face ID';
@@ -7,10 +12,11 @@ const withLocalAuthentication: ConfigPlugin<{ faceIDPermission?: string } | void
   config,
   { faceIDPermission } = {}
 ) => {
-  if (!config.ios) config.ios = {};
-  if (!config.ios.infoPlist) config.ios.infoPlist = {};
-  config.ios.infoPlist.NSFaceIDUsageDescription =
-    faceIDPermission || config.ios.infoPlist.NSFaceIDUsageDescription || FACE_ID_USAGE;
+  withInfoPlist(config, (config) => {
+    config.modResults.NSFaceIDUsageDescription =
+      faceIDPermission || config.modResults.NSFaceIDUsageDescription || FACE_ID_USAGE;
+    return config;
+  });
 
   return AndroidConfig.Permissions.withPermissions(config, [
     'android.permission.USE_BIOMETRIC',