+++ packages/expo-sensors/plugin/src/withSensors.ts
@@ -1,4 +1,4 @@
-import { ConfigPlugin, createRunOncePlugin } from 'expo/config-plugins';
+import { ConfigPlugin, createRunOncePlugin, withInfoPlist } from 'expo/config-plugins';
 
 const pkg = require('expo-sensors/package.json');
 const MOTION_USAGE = 'Allow $(PRODUCT_NAME) to access your device motion';
@@ -7,10 +7,11 @@ const withSensors: ConfigPlugin<{ motionPermission?: string } | void> = (
   config,
   { motionPermission } = {}
 ) => {
-  if (!config.ios) config.ios = {};
-  if (!config.ios.infoPlist) config.ios.infoPlist = {};
-  config.ios.infoPlist.NSMotionUsageDescription =
-    motionPermission || config.ios.infoPlist.NSMotionUsageDescription || MOTION_USAGE;
+  withInfoPlist(config, (config) => {
+    config.modResults.NSMotionUsageDescription =
+      motionPermission || config.modResults.NSMotionUsageDescription || MOTION_USAGE;
+    return config;
+  });
 
   return config;
 };