+++ packages/expo-local-authentication/CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### 💡 Others
 
+- Prevent config plugin from writing permissions until prebuild.
 - drop unused web `name` property. ([#27437](https://github.com/expo/expo/pull/27437) by [@EvanBacon](https://github.com/EvanBacon))
 
 ## 13.8.0 — 2023-11-14