+++ packages/expo-sensors/CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### 💡 Others
 
+- Prevent config plugin from writing permissions until prebuild.
 - drop unused web `name` property. ([#27437](https://github.com/expo/expo/pull/27437) by [@EvanBacon](https://github.com/EvanBacon))
 - Fixed broken JavaScript unit tests. ([#27257](https://github.com/expo/expo/pull/27257) by [@kudo](https://github.com/kudo))